Player HP를 float형으로 변경

+++ Hot Air Balloon/Assets/Scripts/Player.cs
@@ -9,8 +9,8 @@ public class Player : MonoBehaviour
 
     public Text lifeText;
 
-    public int maxHP = 100;
-    public int currentHP;
+    public float maxHP = 100;
+    public float currentHP;
     public int life = 1;
     public int level = 1;
